+++ api/src/helper.hpp
@@ -16,9 +16,9 @@
     #define MAC
 #endif
 
-void sha256(const char* in, uint64_t in_len, char out[65]);
-char* deflate(const char* in, uint64_t in_len, uint64_t& out_len);
-char* inflate(const std::string& in, uint64_t& out_len);
+void sha256(const char* in, size_t in_len, char out[65]);
+char* deflate(const char* in, size_t in_len, size_t& out_len);
+char* inflate(const std::string& in, size_t& out_len);
 
 std::vector<std::string> path_from_input(std::filesystem::path repository_root, const std::vector<std::string>& inputs, const std::string& version_id, LvcError& err);
 void insert_pattern(std::string& content, const std::string& type);
@@ -27,8 +27,8 @@ std::filesystem::path workspace_path(const std::filesystem::path& workspace_dir,
 
 namespace io {
     LvcError file(std::filesystem::path path, std::ios_base::openmode flags);
-    LvcError file(std::filesystem::path path, std::ios_base::openmode flags, const char* content, uint64_t length, bool compress);
-    LvcError prefix_file_content(std::filesystem::path path, const char* content, uint64_t length);
+    LvcError file(std::filesystem::path path, std::ios_base::openmode flags, const char* content, size_t length, bool compress);
+    LvcError prefix_file_content(std::filesystem::path path, const char* content, size_t length);
     bool dir(std::filesystem::path lvc);
     std::string content(std::filesystem::path file_path, bool decompress);
     std::string content_first_line(std::filesystem::path file_path);
