+++ api/src/helper/insert_pattern.cpp
@@ -1,9 +1,9 @@
 #include <helper.hpp>
 
 void insert_pattern(std::string& content, const std::string& type) {
-    uint64_t content_size = content.size();
+    size_t content_size      = content.size();
     const std::string header = type + ' ' + std::to_string(content_size) + '\n';
-    uint64_t total_len = header.size() + content_size;
+    size_t total_len         = header.size() + content_size;
 
     content.resize(total_len);
     if (content_size)
